[Infra] move CoreDataFeedStore and Managed Object to production
