+++ FeedStoreChallenge/CoreDataFeedStore.swift
@@ -1,0 +1,39 @@
+//
+//  CoreDataFeedStore.swift
+//  FeedStoreChallenge
+//
+//  Created by Chad Chang on 2021/2/21.
+//  Copyright © 2021 Essential Developer. All rights reserved.
+//
+
+import CoreData
+
+public final class CoreDataFeedStore: FeedStore {
+	public init() {
+		
+	}
+	public func deleteCachedFeed(completion: @escaping DeletionCompletion) {
+
+	}
+
+	public func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
+
+	}
+
+	public func retrieve(completion: @escaping RetrievalCompletion) {
+		completion(.empty)
+	}
+}
+
+private class ManagedCache: NSManagedObject {
+	@NSManaged var timestamp: Date
+	@NSManaged var feed: NSOrderedSet
+}
+
+private class ManagedFeedImage: NSManagedObject {
+	@NSManaged var id: UUID
+	@NSManaged var imageDescription: String?
+	@NSManaged var location: String?
+	@NSManaged var url: URL
+	@NSManaged var cache: ManagedCache
+}
